Вместо использования subprocess, был перенесен код update.py внутрь check_updates, чтобы избежать дублирования окон при запуске сабпроцесса. Файл update.py был удалён.
