--- config.ini
@@ -1,5 +1,0 @@
-[paths]
-local_minecraft =
-tlauncher =
-local_mods =
-
