--- mods_to_update.json
@@ -1,1 +1,0 @@
-[]
